pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "myss-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlmodel>=0.0.19",
    "alembic>=1.13.0",
    "redis>=5.0.4",
    "structlog>=24.1.0",
    "tenacity>=8.3.0",
    "pybreaker>=1.2.0",
    "celery>=5.4.0",
    "httpx>=0.27.0",
    "PyJWT>=2.8.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    "bcrypt>=4.0.0",
    "pydantic-settings>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "respx>=0.21.0",
    "fakeredis>=2.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
